requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
langchain
langchain-openai
langchain-community
langchain-huggingface
langgraph
faiss-cpu
pypdf
unstructured
markdown
sentence-transformers
